+++ AppleWinExpress2019.vcxproj
@@ -109,7 +109,7 @@
     <ClInclude Include="source\Tfe\Uilib.h" />
     <ClInclude Include="source\Video.h" />
     <ClInclude Include="source\Windows\DirectInput.h" />
-    <ClInclude Include="source\Windows\DirectVideo.h" />
+    <ClInclude Include="source\Windows\WinVideo.h" />
     <ClInclude Include="source\YamlHelper.h" />
     <ClInclude Include="source\z80emu.h" />
     <ClInclude Include="source\Z80VICE\daa.h" />
@@ -215,7 +215,7 @@
     </ClCompile>
     <ClCompile Include="source\Video.cpp" />
     <ClCompile Include="source\Windows\DirectInput.cpp" />
-    <ClCompile Include="source\Windows\DirectVideo.cpp" />
+    <ClCompile Include="source\Windows\WinVideo.cpp" />
     <ClCompile Include="source\YamlHelper.cpp" />
     <ClCompile Include="source\z80emu.cpp" />
     <ClCompile Include="source\Z80VICE\daa.cpp">
