+++ source/AppleWin.cpp
@@ -54,7 +54,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Speech.h"
 #endif
 #include "SynchronousEventManager.h"
-#include "Windows/DirectVideo.h"
+#include "Windows/WinVideo.h"
 #include "RGBMonitor.h"
 #include "NTSC.h"
