+++ source/Frame.cpp
@@ -52,7 +52,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #ifdef USE_SPEECH_API
 #include "Speech.h"
 #endif
-#include "Windows/DirectVideo.h"
+#include "Windows/WinVideo.h"
 
 #include "../resource/resource.h"
 #include "Configuration/PropertySheet.h"
