+++ source/Windows/WinVideo.cpp
@@ -28,7 +28,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "StdAfx.h"
 
-#include "Windows/DirectVideo.h"
+#include "Windows/WinVideo.h"
 #include "AppleWin.h"
 #include "Video.h"
 #include "CPU.h"
