+++ source/Windows/WinVideo.h
